show all entries on song main page + layout fix for "per album" toggle
